add dynamic team detail sidebar
